Using music21 types explicitly
Unified music21 types in single module.

Had to use music21 types eimported implicitly and the fact that they
were imported each from separate module was a little bloaty, so made
a little aggrate module to keep them together.

The goal of this change is to allow using names like Note for own code
in future.